feat(account): implement account transaction history
Added new functionalities to the `Account` microservice. Now, every account balance change triggers an event that is captured and saved as a transaction history. Added `AccountTransaction` entity and `AccountTransactionService` to handle this operation. Also, added `spring-modulith` dependencies to the `pom.xml` file for better modularity. Commented out the test cases in `AccountServiceTest` for further modification. Made the `Account` class and `AccountRepository` interface public for access from the new transaction service.