test(accounts_service): remove balance-related tests and update existing ones
The balance-related tests in the `AccountServiceTest` and `AccountIntegrationTest` classes have been removed. This is due to the removal of the balance property in the `Account` entity.

The existing tests in these classes have also been updated to reflect this change. In particular, the `AccountServiceTest` no longer mocks the `ApplicationEventPublisher` as it's no longer used in the `AccountServiceImpl`. The `AccountIntegrationTest` has been updated to use the `UpdateAccountBalanceDto` for updating account balance and to test getting balance in PLN currency.

These changes are part of refactoring the balance handling logic in the accounts service.