feat(EXCHANGE-SERVICE): add initial configuration and main application file
This commit includes the set up of the initial configuration for the `EXCHANGE-SERVICE` microservice. The `application.yml` and `application-dev.yml` files have been added to define the application specific properties and settings for the dev profile respectively. A new `ExchangeServiceApplication.java` file has also been added as the main application class.