feat(card-service): add new card service microservice
This commit includes the initial setup and configuration for the new card service microservice. It includes new Java application files, application configurations, Maven wrapper, Maven configuration, and a new H2 database setup. The microservice is setup to run on port 8086 in development mode. It also includes settings for Eureka client for service discovery.