feat(card-service): add Postgres and Docker support
Add PostgreSQL and Docker support to the card-service module. This is achieved by adding necessary dependencies to `pom.xml` and creating Docker compose files. Also, reconfigured `application.yml`, `application-dev.yml` and `application-prod.yml` files to accommodate the new changes. Now, in the production environment, the application will connect to a Postgres database running in a Docker container. In the development environment, it will connect to an H2 in-memory database. The Docker compose file is configured to only run during the production profile.