feat(notification-service): add docker integration and restructure configuration
This commit introduces a new production profile for the `notification-service` and restructures the application configuration files. The new production profile includes a Docker Compose setup to spin up a Postgres database as a service.

Changes include:

- Addition of `application-prod.yml` and `compose.yml` for production Docker Compose setup.
- Addition of dependencies in `pom.xml` for Docker Compose and Postgres.
- Restructuring of `application-dev.yml` to only include development specific configurations.
- Updates to `application.yml` to include shared configurations applicable to both the development and production profiles.

These changes allow for a clear separation of development and production configurations, and make deployment and scaling in a production environment much simpler.