refactor(exchange-service): Improve code readability and configuration settings
Refactor `CurrencyHttpClient` in the `exchange-service` by enhancing the control flow to improve code readability. Removed unnecessary `continue` statement and encapsulated the currency retrieval logic inside an `else` block.

Configuration settings in `application-prod.yml` and `application-dev.yml` files are updated as well. Unnecessary comments and empty lines were removed and the `add-properties` configuration is set to `false` in the prod. In dev, `h2.console.enabled` is set to `true` for better debugging.

Slight modification in `CurrencyScheduler` by removing an unnecessary empty line for cleanliness.

These changes are intended to make the code more readable and maintainable, and to optimize the application configurations in different environments.